# swagger.yaml
openapi: 3.0.3
info:
  title: Today Farm API
  description: 오늘농장의 API
  version: 0.1.0
servers:
  - url: http://34.64.88.23
...
paths:
...
  /devices:
    get:
      tags:
        - DEVICE
      description: 회원의 디바이스 목록 조회
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    deviceId:
                      type: integer
                      example: 100
                    name:
                      type: string
                      example: "상도의 상추"
                    status:
                      type: integer
                      example: 1
        "400":
          description: Bad Request
    post:
      tags:
        - DEVICE
      description: 디바이스 추가
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Device"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Device"
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    deviceId:
                      type: integer
                      example: 100
                    name:
                      type: string
                      example: 상도의 상추
        "400":
          description: Bad Request
    put:
      tags:
        - DEVICE
      description: 디바이스 삭제
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceId:
                  type: integer
                  example: 100
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                deviceId:
                  type: integer
                  example: 100
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  deviceId:
                    type: integer
                    example: 100
                  name:
                    type: string
                    example: "상도의 상추"
        "400":
          description: Bad Request
...
components:
  schemas:
...
    Device:
      type: object
      required:
        - deviceId
        - name
      properties:
        deviceId:
          type: integer
          example: 100
        name:
          type: string
          example: 상도의 상추
...